rearrange folder structure, change table rendering method, set sticky top bar
