+++ src/app/page.jsx
@@ -1,6 +1,4 @@
 'use client';
-import TableView from './components/TableView';
-import CardView from './components/CardView';
 import { ViewSwitchToggle } from './components/ViewSwitchToggle';
 import React from 'react';
 import { observer } from 'mobx-react';
@@ -9,6 +7,8 @@ import SearchBar from './components/SearchBar';
 import Filter from './components/Filter';
 import { Box, CircularProgress, createTheme, CssBaseline, ThemeProvider } from '@mui/material';
 import { ThemeSwitchButton } from './components/ThemeSwitchButton';
+import TableView from './pages/table/TableView';
+import CardView from './pages/card/CardView';
 
 const Home = observer(() => {
   const theme = createTheme({
@@ -27,10 +27,10 @@ const Home = observer(() => {
     <ThemeProvider theme={theme}>
       <CssBaseline />
       {/* Outer container */}
-      <div className='flex flex-col h-screen w-full overflow-y-scroll'>
+      <div className='flex flex-col w-full'>
 
         {/* Header container */}
-        <div className='mx-[10rem] py-4'>
+        <div className='mx-[10rem] py-4 sticky top-0 z-10'>
           <div className='flex justify-between items-center'>
             <h1>Countries {pageModeStore.tableView ? 'Table' : 'Card'} View</h1>
             <ThemeSwitchButton />
@@ -40,7 +40,7 @@ const Home = observer(() => {
         {/* Main container */}
         <div className='flex flex-col flex-grow bg-gray-100'>
           {/* Search bar, filter, and view switch */}
-          <div className='flex justify-between items-center mx-[11rem] py-4'>
+          <div className='flex justify-between items-center mx-[11rem] py-4 sticky top-[4rem] z-10 bg-gray-100'>
             <SearchBar />
             <div className='flex items-center space-x-4'>
               <Filter />
